add confirm to dele, add update command, add color to logs

+++ internal/cmd/delete.go
@@ -1,7 +1,9 @@
 package cmd
 
 import (
+	"fmt"
 	"log/slog"
+	"strings"
 
 	"github.com/hopstee/sshvault/internal/storage"
 	"github.com/spf13/cobra"
@@ -21,14 +23,27 @@ func (c *Command) deleteCmd() {
 				slog.Error("connection not found")
 				return
 			}
+
+			fmt.Printf("Delete %s connection? (Y/n): ", conn.Name)
+			var deleteRes string
+			_, err = fmt.Scanln(&deleteRes)
+			if err != nil {
+				slog.Error("failed to read your decision", slog.Any("error", err))
+				return
+			}
+			if strings.ToLower(deleteRes) == "n" {
+				slog.Info("deletion rejected")
+				return
+			}
+
 			if err := c.storage.Delete(name); err != nil {
-				slog.Error("failed to delete connection from store", "err", err)
+				slog.Error("failed to delete connection from store", slog.Any("error", err))
 				return
 			}
 
 			if conn.AuthType == storage.PasswordAuth && conn.PasswordKey != "" {
 				if err := c.keyring.Delete(conn.PasswordKey); err != nil {
-					slog.Error("failed to delete connection from keyring", "err", err)
+					slog.Error("failed to delete connection from keyring", slog.Any("error", err))
 					return
 				}
 			}
